audit/requirements.txt
django>=5.0
openai
slither-analyzer
web3
PyJWT